pyproject.toml
[tool.poetry]
name = "socnet"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
autopep8 = "^2.0.2"
pylint = "^2.17.4"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"
aiologger = "^0.7.0"
redis-purse = "^1.1.0"
async-fastapi-jwt-auth = "^0.5.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"